requirements.txt
flask==2.0.1
pymongo==4.1.1
requests==2.26.0
python-dotenv==0.19.0
google-generativeai==0.3.0
psutil==5.8.0
gunicorn==20.1.0 